+++ src/main/java/com/hust/baseweb/applications/logistics/service/ProductPriceService.java
@@ -4,12 +4,14 @@ import com.hust.baseweb.applications.logistics.entity.ProductPrice;
 import com.hust.baseweb.entity.UserLogin;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-
 
 @Service
 public interface ProductPriceService {
-    ProductPrice setProductPrice(UserLogin createdByUserLogin, String productId, BigDecimal price, String currencyUomId, String taxInPrice);
+    ProductPrice setProductPrice(UserLogin createdByUserLogin,
+                                 String productId,
+                                 Double price,
+                                 String currencyUomId,
+                                 String taxInPrice);
 
     ProductPrice getProductPrice(String productId);
 }
